requirements.txt
Django
psycopg2
Pillow
python-dotenv
djangorestframework 
markdown 
django-filter
qrcode